Fix TODO in calculateLevelFound
A solution to the TODO in calculateLevelFound, respecting the level step settings in creatureCollection.
Getting access to the creatureCollection requires many places to change, passing the collection during creature creation, during any stat recalculation (breeding planner, etc). It's either that or static access to the Form1's creatureCollection. Your choice.