Hopefully fix the updater for the zip version of ASB
Removed code that inherently will **always** throw exceptions.
Exception was related to attempting to extract a ZIP archive to an existing directory, specifically the Current Directory.
Minimal testing was done, but appears to actually work now.

TODO: The updater is moved to the temp directory to complete it's task, but is not removed afterwards.