Modified growingUntil calc for imported saves (#1230)
Fixes a bug where babies that have been cryoed will never display the growing until time, unless you manually set it. This occurred because after being cryoed the TamedAtTime never gets set when the dino is deployed again.

Also has the side effect of showing when the dino would be fully grown if taken out of a cryopod immediately.